V1.2: UI Changes, New Setting & Search QOL
Changed the scale of some text elements slightly.
Made it so when a new entry is added via the searchbox it is appended into the results list.
Added a new setting 'edit_popups', if enabled the app functions as before but if disabled you will no longer get popups when trying to edit the name, login or password on an entry.
Made it so when opening the app if user settings don't have a key:value pair present in the default user data it will add it to your user data, allowing upgrading into versions with new settings without having to manually edit your user data, this process only runs if the data was not empty and therefor set to the default.
Because of these and recent changes prior the version in the app has been updated to 1.2.