Fixed Bugs & Remove Unused Nodes
Fixed bugs in the SearchElement when trying to search through data that doesn't exist.
Removed unused timer nodes in scenes and made PageSelector delete the container it isn't using.
When searching the SearchElement now tells you where in the search process it is, and if there are no results says custom 'No Results Found' message opposed to '0 Results Found'
Updated version in the project settings to 1.2 to match real version.