Fixed UI & Infopanel, Removed Exported Web Build.
Fixed issues in the UI and Infopanel, including buttons not having their mouse cursors set and Infopanel having issues on changing the name of Tag entires.
A Web version is no longer exported by default, the preset still exists to allow for exporting but is no longer included in the Exported Project list.
Minorly optimised the Search Engine by removing an unnecessary variable (special_search) alongside slightly scaling down the search result string in the UI to stop it shifting the interface around.