1.2.1: Updated to Godot 4.4.1 and made release compatible with NSL.
Updated from Godot 4.3 > 4.4.1

Updated ReadMe to mention the releases page on the GitHub.

The main purpose of this update is to allow for a new release to be compatible with the new NSL (NatZombieGames Software Launcher), as that relies on standards I had not yet implemented when I released this project initially.